src/requirements.txt
auto_mix_prep==0.2.0
flask==2.1.2
numpy==1.23.1
pandas==1.4.3
plotly==5.8.2
PyYAML==6.0
Werkzeug==2.1.2
gunicorn==20.1.0
